requirements.txt
akshare>=1.10.0
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
pyyaml>=6.0
requests>=2.28.0
